Dedupe e_spear_guard
