+++ config/symbols.us.stno4.txt
@@ -7,6 +7,7 @@ g_EInitSpawner = 0x80180BB0;
 g_EInitInteractable = 0x80180BBC;
 g_EInitCommon = 0x80180BE0;
 g_EInitDamageNum = 0x80180BEC;
+g_EInitSpearGuard = 0x80180C88;
 NO4_RedDoorTiles = 0x80180D2C;
 g_SplashSfx = 0x80180EE4;
 g_SurfacingSfx = 0x80180EE6;
@@ -160,4 +161,3 @@ g_LayoutObjVertical = 0x801DEB74;
 g_LayoutObjPosHorizontal = 0x801DEB78;
 g_LayoutObjPosVertical = 0x801DEB7C;
 g_ItemIconSlots = 0x801DF81C;
-g_EInitSpearGuard = 0x80180C88;
