+++ src/st/no4/no4.h
@@ -64,8 +64,8 @@ typedef enum EntityIDs {
     // /* 0x37 */ E_ID_37,
     // /* 0x38 */ E_ID_38,
     // /* 0x39 */ E_ID_39,
-    /* 0x3A */ E_ID_3A = 0x3A,
-    /* 0x3B */ E_ID_3B,
+    /* 0x3A */ E_UNK_NO4_SPEAR_GUARD = 0x3A,
+    /* 0x3B */ E_SPEAR_GUARD_BLOCK,
     // /* 0x3C */ E_ID_3C,
     // /* 0x3D */ E_ID_3D,
     // /* 0x3E */ E_ID_3E,
